+++ modules/School_Setup/Periods.php
@@ -132,8 +132,15 @@ if ( ! $_REQUEST['modfunc'] )
 		'LENGTH' => '_makeTextInput',
 	] ); //	'ATTENDANCE' => '_makeCheckboxInput','START_TIME' => '_makeTimeInput','END_TIME' => '_makeTimeInput'
 
-	$columns = [
-		'REMOVE' => '<span class="a11y-hidden">' . _( 'Delete' ) . '</span>',
+	$columns = [];
+
+	if ( empty( $_REQUEST['LO_save'] ) )
+	{
+		// Do not Export Delete column.
+		$columns['REMOVE'] = '<span class="a11y-hidden">' . _( 'Delete' ) . '</span>';
+	}
+
+	$columns += [
 		'TITLE' => _( 'Title' ),
 		'SHORT_NAME' => _( 'Short Name' ),
 		'SORT_ORDER' => _( 'Sort Order' ),